im done with this shit

+++ app/controllers/parties_controller.rb
@@ -10,24 +10,29 @@ class PartiesController < ApplicationController
 
   def new
   	@parties = Party.new
+    redirect_to root_path unless @current_user && @current_user.admin?
   end
 
   def create
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@party = Party.create party_params
   	redirect_to party_path(@party)
   end
 
   def edit
   	@party = Party.find params[:id]
+    redirect_to root_path unless @current_user && @current_user.admin?
   end
 
   def destroy
+    redirect_to root_path unless @current_user && @current_user.admin?
     party = Party.find params[:id]
     party.destroy
     redirect_to parties_path
   end
 
   def update
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@party = Party.find params[:id]
   	if @party.update party_params
   		redirect_to root_path
