+++ app/controllers/politicians_controller.rb
@@ -8,25 +8,30 @@ class PoliticiansController < ApplicationController
   end
 
   def new
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@politicians = Politician.new
   end
 
   def create
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@politician = Politician.create politician_params
   	redirect_to politician_path(@politician)
   end
 
   def edit
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@politician = Politician.find params[:id]
   end
 
   def destroy
+    redirect_to root_path unless @current_user && @current_user.admin?
     politician = Politician.find params[:id]
     politician.destroy
     redirect_to politicians_path
   end
 
   def update
+    redirect_to root_path unless @current_user && @current_user.admin?
   	@politician = Politician.find params[:id]
   	if @politician.update politician_params
   		redirect_to root_path
